requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
requests>=2.32,<3.0
beautifulsoup4>=4.12,<5.0
jinja2>=3.1,<4.0
python-dotenv>=1.0,<2.0
pydantic>=2.9,<3.0
playwright>=1.58,<2.0
